interviewready:gameengine:TTT:GameTest incomplete, the business logic has bugs

+++ src/case_studies/interviewready/ai_game_engine/game/Player.java
@@ -31,11 +31,11 @@ public class Player {
         return "Player-'" + playerSymbol + "' ";
     }
 
-    public void setTimeTaken(int timeInMillis) {
+    public void setTimeTaken(long timeInMillis) {
         timeUsedInMillis += timeInMillis;
     }
 
-    public int getTimeUsedInMillis() {
+    public long getTimeUsedInMillis() {
         return this.timeUsedInMillis;
     }
 }
